created ClearAllButton component and moved clear button functionality here
