+++ code/src/components/ClearAllButton.js
@@ -6,12 +6,11 @@ import swal from "sweetalert";
 import { todos } from "reducers/todos";
 
 export const ClearAllButton = () => {
-  // const items = useSelector(store => store.todos.items);
   const dispatch = useDispatch();
 
   const handleOnDelete = () => {
     swal({
-      title: "Delete all To-Do's?",
+      title: "Clear all To-Do's?",
       text: "Are you sure you want to delete all of your To-Do's?",
       icon: "warning",
       dangerMode: true,
@@ -19,9 +18,9 @@ export const ClearAllButton = () => {
     }).then(willDelete => {
       if (willDelete) {
         dispatch(todos.actions.removeAll());
-        swal("Deleted!", "Your To-Do's have been deleted", "warning");
+        swal("Cleared!", "Your To-Do's have been deleted!", "warning");
       } else {
-        swal("Cancelled", "Your To-Do's are still a work in progress!");
+        swal("Cancelled!", "Your To-Do's are still a work in progress!");
       }
     });
   };
